updated demo project, crated subscription sample with documentation
